# Faster /rank